add optional semicolon handling
This required a significant refactoring of the lexer, in order ot track
optional semicolons everywhere they might be inserted, and handle them
appropriately.

We change the way the lexeme and symbol combinators work to, instead of
consuming trailing whitespace, consuming *leading* whitespace. This
allows us to not inadvertantly consume precious newline characters that
we need to handle specially. Then, for all parsers that could parse
constructs that may imply a semicolon in Go, we change their type from
`Parser a` to `Parser (Semi a)`. The Semi monad is a state monad
transformer stacked onto an exception monad. The state type is `Maybe
Bool` and tracks whether a semicolon detection has been performed (this
is what the Maybe wrapper is for) and if so, whether there is a
semicolon next, explicit or otherwise. The exception type represents
failures such as there not having been a semicolon detection performed
or unexpected semicolons present in the input.

We introduce the `withDetectSemicolon` combinator to run a given parser
and check for semicolons after, either by matching an explicit semicolon
or by reaching the end of the line. This combinator introduces a new
`Semi` computation.

The functions `noSemi`, `requireSemi`, and `unSemi` allow us to get out
of the Semi monad by requiring that there be no semicolon, that there be
a semicolon, or that we don't care about whether there's a semicolon,
respectively. Some of these combinators have a variety ending in `P`
that reinterpret the Semi monad's result in the Parser monad, which is
useful to discharge Semi monad computations in the context of a Parser.

For example, `noSemiP :: Semi a -> Parser a` requires that there be no
semicolon in the context of a parser. If the Semi computation failed
with an exception, then that exception is reinterpreted as a parsing
error in the Parser monad; else, the result of the Semi is simply
returned in the Parser monad.

Many tests needed to be corrected in order to correctly typecheck
(unwrapping `Semi` monads).