create pretty-print framework and implement expression pretty-print
We introduce the typeclass `Pretty a` of types that can be
pretty-printed. It uses the exact same structure as the Show typeclass,
but has different semantics. Specifically, rather than create output
that is meant to be valid Haskell code (this is a goal of Show), Pretty
aims to produce output that is valid Go.

Simple expressions, consisting of binary and unary operators, literals,
and variables can be pretty-printed. More complicated expressions are
not supported (yet).