Fixed conflict between float literal and selector.
The `expr` parser wasn't able to tell if `3.` was the beginning of
a float literal or a selector expression. This resulted in it being
unable to parse a float literal properly. The `floatLiteral` parser
was changed to add a `notFollowedBy identifier` after the dot.

Also switched the order of `floatLiteral` and `intLiteral` so that
`literal` would try the former first, because otherwise it can't
parse something like `1.2` properly.