Added support for increment/decrement statements (++, --).
Also made sure that conflicts between the assignment operators and
others were handled properly. Trying to parse an expression
containing an assignment operator will fail, and trying to parse
an assignment statement will work as intended (right now we can
even parse something like `x + x++`, which is a side-effect of
allowing any expression as the operand - this will be fixed).

Also added some documentation to the simple statement parsers.