Added parser for type declarations.
This required creating a parser for struct types. Because those can
appear in a variety of contexts (essentially anywhere another type
can appear, such as in a variable declaration or a function definition),
the syntax tree was reworked slightly to create an appropriate
StructType constructor under the Type data type. StructType is a
constructor with one argument, a list of tuples of lists of identifiers
and types. The tuple represents a field declaration of the structure.

Thanks to this change, the data types for type declarations and variable
declarations are now much more similar (there is no need for a separate
StructDecl constructor in TypeDecl), so their parsers are also nearly
exactly the same at a high level (only the keyword used and the
specification parser differ).

A case was also added to the pretty instance of Type to pretty-print
StructTypes.