simplify definition of PrintStmt
We can at parse time look whether a print statement is a print or a println and
inject a synthetic "\n" expression on the end of the expressions to print
accordingly. As a result, we don't need to track in the internal representation
whether print or println was used specifically.