improve pretty printer
We add a derived instance for StatementF in order to use a catamorphism to
pretty-print the recursive structure.

We finish the Pretty instance for Fix ExprF. The F-algebra used by the
catamorphism to pretty-print Fix ExprF is now total.

We add total Pretty instances for the following types: TopLevelDecl, FunDecl

We add a partial Pretty instance for Fix StatementF implementing the following
cases:

* DeclStmt
* ExprStmt
* ShortVarDecl
* Assignment
* PrintStmt
* ReturnStmt
* BreakStmt
* ContinueStmt
* FallthroughStmt