Various fixes and tests for `switch`.
One of the failing tests for `switch` involved the fact that any
expression was considered a simple statement. This was fixed by
adding a check to the relevant parser. To do this, a new function
was introduced in Semi: `condUnSemiP`, which works much like
`unSemiP` except that it also takes a predicate to apply on the
parsed value, and an error message. If the predicate fails, the
failure parser is invoked with the given error message. This gives
us good error messages for any condition we want to check on `Semi`
results.

Closes #5 (finally!)