Added classification for integral operators (closes #50).
Further, implemented assign-op typechecking in terms of binary op
typechecking. We don't want inconsistencies between the two.

Also changed the return types of some built-ins. They were previously
unknown in the case of an error, but for some of them we still know
what the type of the call should be.