rethink how canonical types are used and improve FuncDecl representation
We introduce the `TySrcAnnType` type synonym that is a fixed point in which the
branches are annotated with type and source position information but for which
the base functor is just `SrcAnnTypeF`.

The FuncDecl data type had a `ty` parameter used to represent types, but it
turned out to be useful to separate the argument types from the return type.