perform struct field lookup on bare Ident
The Prelude.lookup function uses the Eq instance of the key to perform a lookup
in an association list. The Eq instance for SrcAnnIdent is the derived
instance, which will compare the annotations for equality as well. Needless to
say, this kind of lookup will always fail, as it's impossible to construct a
selector identifier whose source position is the same as the identifier used to
define the field in the struct.

Closes #83