refactor GlobalId and reorganize GoLite.Types
GlobalId is moved to its own module and made abstract over the type of the
identified object as well as its original representation. This will allow
GoLite and Vigil to use different data in those locations.

An explicit export list is added to GoLite.Types with proper headers to
organize the generated haddocks.