Closes #103.
TyAnn was restored. We need to know the types of the temporaries
we're generating (since they're global IDs), and so we need to
know the types of the expressions/refs we're generating them for.
This caused a huge amount of boilerplate, but now we know what the
types of our temporaries are.

Also re-added void type, since we need it to represent function
types and their returns.