ensure 16 byte stack alignment and refactor
Add a useless push/pop when saving/loading scratch registers around function
calls when the number of pushes/pops is not a multiple of two. On function
entry/exit, round up the stack space allocated to a multiple of 16.

Remove stack parameter for register allocation now that the compilation phase
interacts only with virtual registers and not with the stack.